lt-memo/requirements.txt
strands-agents[bedrock]
bedrock-agentcore
bedrock-agentcore-starter-toolkit
boto3
